==> Notebooks/NB1-City_Private Spreadsheet.ipynb <==
### Cleaning So. Cal Soccer Activity & Rent Data

---

Summary of Data Collection (before cleaning):
- Created spreadsheets of active soccer leagues (city-run and private) and building rent data in LA & Orange counties

---

import pandas as pd
import geopandas as gpd
import googlemaps
from shapely.geometry import Point
import numpy as np
import array

---

#### Cleaning Spreadsheet Data:

oc = pd.read_csv('OC_Soccer_Fields1.csv')

la = pd.read_csv('LA_Soccer_Fields1.csv')

oc.head(1)

la.head(1)

leagues_df = pd.concat([oc, la], sort=False).reset_index(drop=True)

# Creating separate columns for name & address
leagues_df[['Private1', 'Private1_Address']] = leagues_df['Private League1'].str.split('; ', expand=True)
leagues_df[['Private2', 'Private2_Address']] = leagues_df['Private League2'].str.split('; ', expand=True)

leagues_df = leagues_df.drop(['Private League1', 'Private League2'], axis='columns').fillna('No')

# Changing the names of these cities due to different naming in data used in later notebooks
leagues_df.loc[leagues_df['City'].str.contains('Industry'), 'City'] = 'Industry'
leagues_df.loc[leagues_df['City'].str.contains('Palos Verdes'), 'City'] = 'Palos Verdes'
leagues_df.loc[leagues_df['City'].str.contains('La Cañada Flintridge'), 'City'] = 'La Canada Flintridge'

leagues_df.head(2)

---

Converting Rent Cols to Floats:

# Since I'm removing the $ via indexing, I replaced 'No' w/a 2 digit number (in string format) so that removing the first digit would result in '0'
rent_cols = ['Avg Rent - Office', 'Avg Rent - Industrial', 'Avg Rent - Retail']

for col in rent_cols:
    leagues_df[col] = leagues_df[col].replace('No', '10').apply(lambda x:x[1:]).astype("float")  

leagues_df.head(2)

---

Adding Activity Column:

cities = []
activity_levels = []

for i, row in leagues_df.iterrows():
    value = 0
    if row.City_Address != 'No':
        value = 2
    if row.Meetup != 'No':
        value += 1
    cities.append(row.City)
    activity_levels.append(value)

leagues_df['Soccer_Activity'] = activity_levels

leagues_df.head(1)

---

#### Adding Lat/Lon Coordinates of Private Leagues:

API_KEY = '(HIDDEN)'

googmap = googlemaps.Client(key=API_KEY)

private1 = leagues_df[['City', 'Private1', 'Private1_Address']][leagues_df['Private1_Address'] != 'No']
private2 = leagues_df[['City', 'Private2', 'Private2_Address']][leagues_df['Private2_Address'] != 'No']

dfs = [private1, private2]

types = ['Private1', 'Private2']
cities = []
names = []
lats = []
lons = []
counter = 0

for df in dfs:
    for i, row in df.iterrows():
        cities.append(row.City)
        names.append(row[types[counter]])
        address = types[counter] + '_Address'
        geocode = googmap.geocode(row[address])
        
        lats.append(float(geocode[0]['geometry']['location']['lat']))
        lons.append(float(geocode[0]['geometry']['location']['lng']))
    counter += 1

private_leagues = pd.DataFrame({'City':cities,'Name':names, 'Lat':lats, 'Lon':lons})

private_leagues.head()

private_leagues.to_csv('private_leagues.csv', index=False)

*This file will be used to plot private league locations as points on a map

---

rent_and_activity_data = leagues_df[['City', 'Avg Rent - Office', 'Avg Rent - Industrial', 'Avg Rent - Retail', 'Soccer_Activity']]

rent_and_activity_data.head()

len(rent_and_activity_data)

rent_and_activity_data.to_csv('rent_and_activity_data.csv', index=False)

*This file will be used in subsequent notebooks to specify which cities to extract census data on

---
