==> Notebooks/NB2-Adding Missing Cities.ipynb <==
#### Merging Rent & Soccer Activity with City Boundary Data

*Adding rows to rent_and_activity_data (from prev. notebook) for missing So. Cal cities- contained in the city boundaries shapefile, below

---

import pandas as pd
import numpy as np
import geopandas as gpd

---

rent_and_activity_df = pd.read_csv('rent_and_activity_data.csv')

rent_and_activity_df.head(2)

rent_and_activity_df.shape[0]

---

# Source: http://gisdata-scag.opendata.arcgis.com/datasets/27b134459761486991f0b72f8a9a67c5_0
cities_shp = 'City_Boundaries_SCAG_Region.shp'

city_boundaries_geodf = gpd.read_file(cities_shp)

city_boundaries_geodf.shape[0]

city_boundaries_geodf.head(1)

---

##### Checking & Dealing w/ Missing Cities:
(Cities not found in city_boundaries)

for city in rent_and_activity_df.City:
    if city not in city_boundaries_geodf.CITY.values:
        print(city)

---

- Check if other variations of PV or WH exist- if so, include them
- Remove 'Palos Verdes' & 'Woodland Hills' after adding variations

---

city_boundaries_geodf[city_boundaries_geodf.CITY.str.contains('Palos')]

city_boundaries_geodf[city_boundaries_geodf.CITY.str.contains('Woodland')].values

---

Replacing name of 'Palos Verdes' & Removing 'Woodland Hills'

rent_and_activity_df[rent_and_activity_df.City == 'Palos Verdes']

rent_and_activity_df.loc[rent_and_activity_df.City == 'Palos Verdes', 'City'] = 'Palos Verdes Estates'

rent_and_activity_df[rent_and_activity_df.City.str.contains('Palos')]

rent_and_activity_df = rent_and_activity_df[rent_and_activity_df.City != 'Woodland Hills']

---

##### Adding Rows for Missing Cities:
*these cities aren't the focus of the analysis but must be included in order to join the shapefile w/ additional statistical data-  each shapefile record must have a corresponding record in the joined statistical data file to successfully join them in QGIS

missing = []
for city in city_boundaries_geodf.CITY:
    if city not in rent_and_activity_df.City.values:
        missing.append(city)

missing_cities_df = pd.DataFrame({'City':missing})

rent_and_activity_df.columns

*Assigning value of 0 for all values in rows of missing cities

for col in rent_and_activity_df.columns[1:]:
    missing_cities_df[col] = np.full(len(missing_cities_df), 0)

missing_cities_df.tail()

all_cities_rent_activity = pd.concat([rent_and_activity_df, missing_cities_df]).reset_index(drop=True)

all_cities_rent_activity.tail()

all_cities_rent_activity.to_csv('all_cities_rent_activity.csv', index=False)

*Will add columns containing census demographic data to this df in the 5th notebook

---
