==> Notebooks/NB3-City PUMAs.ipynb <==
### Linking Cities with PUMA Codes

(PUMA is a US Census geographic unit; this file created in this notebook will be used to link cities to Census demographics in NB4)

---

import pandas as pd

rent_and_activity_data = pd.read_csv('rent_and_activity_data.csv')  # Only includes So Cal cities in LA & Orange counties (from NB1)

len(rent_and_activity_data)

### Examining PUMA Data

# Source: https://usa.ipums.org/usa/volii/cpuma0010.shtml
# PUMA description: https://usa.ipums.org/usa-action/variables/PUMA#description_section
pumas_df = pd.read_csv('PUMA2000_PUMA2010_crosswalk.csv')

pumas_df.head(1)

pumas_df.columns

# The first two PUMA digits correspond to county (which I'll use to filter the data below)
pumas_df['County'] = pumas_df.PUMA10.astype(str).apply(lambda x:x[:2]).astype(int)

puma_cols = ['State10', 'State10_Name', 'PUMA10', 'County', 'PUMA10_Name', 'PUMA10_Pop10', 'PUMA10_Land']

# Including only LA & OC PUMAs
pumas_df = pumas_df[puma_cols][(pumas_df.State10 == 6) & (pumas_df.County.isin([37, 59]))].sort_values('PUMA10')  

# To reveal all data in each cell (instead of cutting off at default max character length)
pd.set_option('display.max_colwidth', -1)

pumas_df.head()

pumas_df = pumas_df.drop_duplicates().reset_index(drop=True)

pumas_df.head()

---

Examining city naming format:

for city_name in pumas_df.PUMA10_Name.apply(lambda x:x.split('--').pop(1)):
    print(city_name)

---

Observations:
- Some cities are located in multiple PUMAs
- Los Angeles City is listed as "LA City"

---

Linking PUMA Codes to Corresponding City Names (in rent_and_activity_data):

len(pumas_df)

len(rent_and_activity_data.City)

# Extracting PUMA(s) corresponding to each city in 'all_rent_activity_data'
city_pumas_dict = {}
missing_pumas = []
counter = 1

for city in rent_and_activity_data.City:
    if city == 'Los Angeles':
        df = pumas_df[pumas_df.PUMA10_Name.str.contains('LA City')]  # To distinguish it from LA County
    elif city == 'Orange':
        df = pumas_df[pumas_df.PUMA10_Name.str.contains('Orange & Villa Park')]  # To distinguish it from Orange County
    else:
        df = pumas_df[pumas_df.PUMA10_Name.str.contains(city)] 
    
    if len(df) == 0:  # i.e. no PUMA corresponding to city in 'combined' df
        missing_pumas.append(city)
        city_pumas_dict[city] = []
    else:
        city_pumas_dict[city] = list(df.PUMA10.values)

len(city_pumas_dict)

city_pumas_dict

# Cities w/ no corresponding PUMA
missing_pumas

# Manually assigned PUMA value of a comparable bordering city that is included in pumas_df 
missing_pumas_dict = {
    'Dana Point':['Laguna Niguel'],
    'La Palma':['Cerritos'],
    'Laguna Beach':['Laguna Niguel'],
    'Laguna Woods':['Laguna Niguel'],
    'Los Alamitos':['Cypress'],
    'Tustin':['Irvine'],
    'Avalon':['Newport Beach'],
    'El Segundo':['Hawthorne'],
    'La Canada Flintridge':['Glendale'],
    'Lomita':['Torrance'],
    'Monrovia':['Arcadia'],
    'San Marino':['San Gabriel']
}

---

Assigning PUMA values to cities w/ missing PUMAs:

for key in missing_pumas_dict:
    pumas = []
    for city in missing_pumas_dict[key]:
        pumas = list(pumas_df.PUMA10[pumas_df.PUMA10_Name.str.contains(city)].values)
    city_pumas_dict[key] = pumas

city_pumas_dict

city_pumas = pd.DataFrame(list(city_pumas_dict.items()), columns=['City', 'PUMAs'])

city_pumas

len(city_pumas)

city_pumas.to_csv('city_pumas.csv', index=False)

---
