==> Notebooks/NB4-Census Demographics.ipynb <==
### Assigning US Census Demographics for Each City

---

import pandas as pd
import numpy as np
import array

city_pumas = pd.read_csv('city_pumas.csv')

len(city_pumas)

city_pumas.head(1)

type(city_pumas['PUMAs'][0])

# Converting lists in str format back to lists containing ints
city_pumas['PUMAs'] = city_pumas['PUMAs'].apply(lambda x:x[1:-1].split(',')).apply(lambda x:[int(s) for s in x])

type(city_pumas['PUMAs'][0])

---

### Cleaning Census Demographic Data

# This Census data was extracted from the following source, by manually choosing the variables of interest: https://usa.ipums.org/usa-action/variables/group
census_df = pd.read_csv('usa_00145.csv')

census_df.head(1)

# County indicated by first two digits of PUMA
census_df['COUNTY'] = census_df.PUMA.astype(str).apply(lambda x:x[:2]).astype(int)

census_df.head(1)

# Keeping only records of individuals residing in LA & Orange counties
census_df = census_df[(census_df.COUNTY.isin([37, 59])) & (census_df.STATEFIP == 6)]

census_df = census_df.drop(['RACED', 'HISPAND'], axis='columns')

census_df.head(1)

---

Cleaning Race Columns:

# Replacing numeric code w/ corresponding race; source: https://usa.ipums.org/usa-action/variables/RACE#codes_section
race = {
1:	'White',
2:	'Black/African American/Negro',
3:	'American Indian or Alaska Native',
4:	'Chinese',
5:	'Japanese',
6:	'Other Asian or Pacific Islander',
7:	'Other race, nec',
8:	'Two major races',
9:	'Three or more major races'}

census_df.RACE = census_df.RACE.map(race)

# Source: https://usa.ipums.org/usa-action/variables/HISPAN#codes_section
hispanic = {
0:	'Not Hispanic',
1:	'Mexican',
2:	'Puerto Rican',
3:	'Cuban',
4:	'Other',
9:	'Not Reported'}

census_df.HISPAN = census_df.HISPAN.map(hispanic)

census_df.RACE.value_counts(normalize=True)

census_df.HISPAN.value_counts(normalize=True)

# Assigning 'Other' to American Indian due to small count; merging Chinese and Japanese with Asian; merging two/three + races and assigning 'Mixed'
races = {'American Indian or Alaska Native': 'Other', 'Black/African American/Negro': 'African American',
       'Other race, nec':'Other', 'White':'White', 'Other Asian or Pacific Islander':'Asian',
       'Two major races':'Mixed', 'Chinese':'Asian', 'Japanese':'Asian',
       'Three or more major races':'Mixed'}

census_df.RACE = census_df.RACE.map(races)

census_df.RACE.value_counts(normalize=True)

census_df.HISPAN.value_counts(normalize=True)

# Simplifying values to 'Hispanic' or 'Not Hispanic'
census_df.loc[census_df.HISPAN != 'Not Hispanic', 'HISPAN'] = 'Hispanic'

census_df.HISPAN.value_counts(normalize=True)

# Reflecting 'Hispanic' in RACE column (which excludes Hispanic race- presumably categorized as 'white' or 'other')
census_df.loc[census_df.HISPAN != 'Not Hispanic', 'RACE'] = 'Hispanic'

census_df.RACE.value_counts(normalize=True)

census_df = census_df.drop('HISPAN', axis='columns')

census_df.head(3)

---

#### Extracting Census Data for Each City

census_df.head(1)

city_pumas.head(1)

cities = []
pumas_called = []
all_incomes = array.array('i')
med_ages = []
med_incomes = []
male_pers = []
hispanic_pers = []
incomes_dict = {}

for city in city_pumas.City:
    pumas = city_pumas.PUMAs[city_pumas.City == city].iloc[0]
    df = census_df[census_df.PUMA.isin(pumas)]
    ages = array.array('i')
    incomes = array.array('i')
    
    for i, row in df.iterrows():
        ages.extend(np.full(row.PERWT, row.AGE))  # Description of PERWT: https://usa.ipums.org/usa-action/variables/PERWT#description_section
        if row.INCTOT != 9999999:  # Code indicating n/a
            incomes.extend(np.full(row.PERWT, row.INCTOT))
        
    # Collecting all incomes w/o overlap (since some cities share the same PUMA)
    for puma in pumas:
        if puma in pumas_called:
            continue
        else:  # Individual PUMA incomes
            for i, row in census_df[census_df.PUMA == puma].iterrows():
                if row.INCTOT != 9999999:  
                    all_incomes.extend(np.full(row.PERWT, row.INCTOT))
                    # Cumulative census incomes
                    if row.INCTOT in incomes_dict:
                        incomes_dict[row.INCTOT] += row.PERWT
                    else:
                        incomes_dict[row.INCTOT] = row.PERWT
            pumas_called.append(puma)
            
    cities.append(city)
    med_ages.append(np.median(ages))
    med_incomes.append(np.median(incomes))
    male_pers.append(round(df.PERWT[df.SEX == 1].sum() / df.PERWT.sum(), 2)) 
    hispanic_pers.append(round(df.PERWT[df.RACE == 'Hispanic'].sum() / df.PERWT.sum(), 2))

len(cities)

---

Comparing Median Incomes:

np.median(all_incomes)  # All recorded incomes in the entire df, weighted by PERWT (which gives the most accurate figure)

np.median(med_incomes)  # Median of all city median incomes

*Conclusion: the median in med_incomes is close enough to all_incomes to rely on

---

demographics_df = pd.DataFrame({'City':cities, 'Median_Age':med_ages, 'Median_Income':med_incomes, 'Normalized_Income':(med_incomes-min(med_incomes)) / (max(med_incomes) - min(med_incomes)), 
                             'Male_Percent':male_pers, 'Hispanic_Percent':hispanic_pers})

demographics_df.Normalized_Income = demographics_df.Normalized_Income.apply(lambda x:round(x, 2))

demographics_df.head()

len(demographics_df)

demographics_df.to_csv('demographics_LA_OC.csv', index=False)

---
