==> Notebooks/NB5-Joining Everything.ipynb <==
### Joining Data From Previous Notebooks
(the two files containing demographic & soccer activity/rent data)

---

import pandas as pd
import numpy as np
import geopandas as gpd

demographics_LA_OC = pd.read_csv('demographics_LA_OC.csv')  # Only includes LA & OC cities

all_rent_activity_data = pd.read_csv('all_rent_activity_data.csv')  # Contains all map cities (including those not in LA or OC)

demographics_LA_OC.tail(1)

all_rent_activity_data.tail(1)

- Add rows w/ cities from all_rent_activity_data that aren't in demographics_LA_OC
- Fill city col w/ cities not in demographics_LA_OC
- Fill empty cells w/ avg. col values
- Merge w/ all_rent_activity_data (to include rent/activity data)

for city in demographics_LA_OC.City:
    if city not in all_rent_activity_data.City.values:
        print(city)

demographics_LA_OC = demographics_LA_OC[~demographics_LA_OC.City.isin(['Palos Verdes', 'Woodland Hills'])]

demographics_LA_OC.shape[0]

missing = []
for city in all_rent_activity_data.City:
    if city not in demographics_LA_OC.City.values:
        missing.append(city)

missing_cities_df = pd.DataFrame(index=range(len(missing)), columns=demographics_LA_OC.columns)

missing_cities_df['City'] = missing

missing_cities_df.head(1)

all_rent_activity_data.tail()

for col in missing_cities_df.columns[1:]:
    missing_cities_df[col] = demographics_LA_OC[col].median()

missing_cities_df.head(1)

combined = pd.concat([demographics_LA_OC, missing_cities_df]).reset_index(drop=True)

combined.tail()

len(combined)

len(all_rent_activity_data)

combined_all = pd.merge(combined, all_rent_activity_data.loc[:,'Avg Rent - Office':], right_index=True, left_index=True)

len(combined_all)

combined_all

combined_all.to_csv('combined_all.csv', index=False)

*The columns in this combined df will be joined w/ the mapping shapefile in QGIS

---
